+++ src/eps/tutorials/ex10.c
@@ -30,7 +30,7 @@ PetscErrorCode STApplyTranspose_User(ST,Vec,Vec);
 PetscErrorCode STApplyHermitianTranspose_User(ST,Vec,Vec);
 #endif
 PetscErrorCode STBackTransform_User(ST,PetscInt,PetscScalar*,PetscScalar*);
-PetscErrorCode STDestroy_User(SampleShellST*);
+PetscErrorCode STDestroy_User(void**);
 
 int main (int argc,char **argv)
 {
@@ -101,6 +101,7 @@ int main (int argc,char **argv)
        this context can be defined to contain any application-specific data. */
     PetscCall(STCreate_User(&shell));
     PetscCall(STShellSetContext(st,shell));
+    PetscCall(STShellSetContextDestroy(st,STDestroy_User));
 
     /* (Required) Set the user-defined routine for applying the operator */
     PetscCall(STShellSetApply(st,STApply_User));
@@ -150,7 +151,6 @@ int main (int argc,char **argv)
     PetscCall(EPSErrorView(eps,EPS_ERROR_RELATIVE,PETSC_VIEWER_STDOUT_WORLD));
     PetscCall(PetscViewerPopFormat(PETSC_VIEWER_STDOUT_WORLD));
   }
-  if (isShell) PetscCall(STDestroy_User(shell));
   PetscCall(EPSDestroy(&eps));
   PetscCall(MatDestroy(&A));
   PetscCall(SlepcFinalize());
@@ -317,10 +317,12 @@ PetscErrorCode STBackTransform_User(ST st,PetscInt n,PetscScalar *eigr,PetscScal
    spectral transformation context.
 
    Input Parameter:
-.  shell - user-defined spectral transformation context
+.  ctx - user-defined spectral transformation context
 */
-PetscErrorCode STDestroy_User(SampleShellST *shell)
+PetscErrorCode STDestroy_User(void **ctx)
 {
+  SampleShellST *shell = (SampleShellST*)*ctx;
+
   PetscFunctionBeginUser;
   PetscCall(KSPDestroy(&shell->ksp));
   PetscCall(PetscFree(shell));