+++ src/sys/classes/st/impls/shell/shell.c
@@ -15,11 +15,12 @@
 #include <slepc/private/stimpl.h>        /*I "slepcst.h" I*/
 
 typedef struct {
-  void           *ctx;                       /* user provided context */
-  PetscErrorCode (*apply)(ST,Vec,Vec);
-  PetscErrorCode (*applytrans)(ST,Vec,Vec);
-  PetscErrorCode (*applyhermtrans)(ST,Vec,Vec);
-  PetscErrorCode (*backtransform)(ST,PetscInt n,PetscScalar*,PetscScalar*);
+  void              *ctx;                       /* user-provided context */
+  PetscCtxDestroyFn *destroy;                   /* context destroy */
+  PetscErrorCode    (*apply)(ST,Vec,Vec);
+  PetscErrorCode    (*applytrans)(ST,Vec,Vec);
+  PetscErrorCode    (*applyhermtrans)(ST,Vec,Vec);
+  PetscErrorCode    (*backtransform)(ST,PetscInt n,PetscScalar*,PetscScalar*);
 } ST_SHELL;
 
 /*@C
@@ -56,7 +57,7 @@ PetscErrorCode STShellGetContext(ST st,void *ctx)
    Logically Collective
 
    Input Parameters:
-+  st - the shell ST
++  st  - the shell `ST`
 -  ctx - the context
 
    Level: advanced
@@ -66,7 +67,7 @@ PetscErrorCode STShellGetContext(ST st,void *ctx)
    for this function that tells Fortran the Fortran derived data type that
    you are passing in as the `ctx` argument.
 
-.seealso: [](ch:st), `STSHELL`, `STShellGetContext()`
+.seealso: [](ch:st), `STSHELL`, `STShellGetContext()`, `STShellSetContextDestroy()`
 @*/
 PetscErrorCode STShellSetContext(ST st,void *ctx)
 {
@@ -76,7 +77,37 @@ PetscErrorCode STShellSetContext(ST st,void *ctx)
   PetscFunctionBegin;
   PetscValidHeaderSpecific(st,ST_CLASSID,1);
   PetscCall(PetscObjectTypeCompare((PetscObject)st,STSHELL,&flg));
-  if (flg) shell->ctx = ctx;
+  if (flg) {
+    if (shell->destroy) PetscCall((*shell->destroy)(&shell->ctx));
+    shell->ctx = ctx;
+  }
+  PetscFunctionReturn(PETSC_SUCCESS);
+}
+
+/*@C
+   STShellSetContextDestroy - Set a context destroy function for the `STSHELL`
+   context.
+
+   Logically Collective
+
+   Input Parameters:
++  st      - the shell `ST`
+-  destroy - context destroy function, see `PetscCtxDestroyFn` for its calling sequence
+
+   Level: advanced
+
+.seealso: [](ch:st), `STSHELL`, STShellSetContext()`
+@*/
+PetscErrorCode STShellSetContextDestroy(ST st,PetscCtxDestroyFn *destroy)
+{
+  ST_SHELL       *shell = (ST_SHELL*)st->data;
+  PetscBool      flg;
+
+  PetscFunctionBegin;
+  PetscValidHeaderSpecific(st,ST_CLASSID,1);
+  PetscAssertPointer(destroy,2);
+  PetscCall(PetscObjectTypeCompare((PetscObject)st,STSHELL,&flg));
+  if (flg) shell->destroy = destroy;
   PetscFunctionReturn(PETSC_SUCCESS);
 }
 
@@ -165,7 +196,10 @@ PetscErrorCode STIsInjective_Shell(ST st,PetscBool* is)
 
 static PetscErrorCode STDestroy_Shell(ST st)
 {
+  ST_SHELL *shell = (ST_SHELL*)st->data;
+
   PetscFunctionBegin;
+  if (shell->destroy) PetscCall((*shell->destroy)(&shell->ctx));
   PetscCall(PetscFree(st->data));
   PetscCall(PetscObjectComposeFunction((PetscObject)st,"STShellSetApply_C",NULL));
   PetscCall(PetscObjectComposeFunction((PetscObject)st,"STShellSetApplyTranspose_C",NULL));